Refactors scenario group diff into their own file
